requirements_bot.txt
# Auth
azure-identity==1.25.2
msal==1.35.0
msal-extensions==1.3.1

# HTTP
httpx==0.28.1
aiohttp>=3.8.0

# Agent Framework
agent-framework==1.0.0rc2
agent-framework-core==1.0.0rc2
agent-framework-azure-ai==1.0.0rc2

# Microsoft Agents (Bot Framework)
microsoft-agents-hosting-aiohttp==0.8.0
microsoft-agents-hosting-core==0.8.0
microsoft-agents-authentication-msal==0.8.0
microsoft-agents-activity==0.8.0

# MCP client (enkel client, niet de server)
mcp==1.26.0

# Config / utils
python-dotenv==1.2.2
pyyaml==6.0.3
anyio==4.12.1